Move capabilities to its own folder
