+++ src/components/ApplicantDetail/component.tsx
@@ -5,7 +5,7 @@ import { DetailHeadline } from "$components//Detail/DetailHeadline";
 import { DetailByLine } from "$components/Detail/DetailByLine";
 import { DetailDescription } from "$components/Detail/DetailDescription";
 import { DetailMainContainer } from "$components/Detail/DetailMainContainer";
-import { CapabilitiesDetail } from "$components/CapabilitiesDetail";
+import { CapabilitiesDetail } from "$components/Capabilities/CapabilitiesDetail";
 import { CareersDetail } from "$components/CareersDetail";
 
 const ApplicantDetail: FunctionComponent<IApplicantDetailProps> = (
