+++ src/components/ApplicantItemsDetail/styles.module.scss
@@ -1,22 +1,24 @@
 @import "$styles/_detailsVariables.scss";
+@import "$components/Capabilities/styles/variables.scss";
 
 .items {
   display: flex;
   flex-direction: column;
-  width: 100%;
+  width: $width;
 
   .item {
-    width: 100%;
-    height: auto;
-    margin-bottom: 5px;
+    width: $width;
+    height: $height;
+    font-size: $font-size;
+    margin-bottom: $margin-bottom;
   }
 
   .title {
     width: 100%;
-    height: auto;
-    font-size: 14px;
+    height: $height;
+    font-size: $font-size;
     font-weight: bold;
-    margin-bottom: 5px;
+    margin-bottom: $margin-bottom;
   }
 }
