+++ src/components/Capabilities/CapabilitiesDetail/component.tsx
