+++ src/components/Capabilities/CapabilitiesDetail/index.tsx
