+++ src/components/Capabilities/CapabilitiesDetail/interface.tsx
