+++ src/components/Capabilities/styles/variables.scss
@@ -1,0 +1,6 @@
+@import "$styles/_detailsVariables.scss";
+
+$font-size: 14px;
+$width: 100%;
+$height: auto;
+$margin-bottom: 5px;
