+++ src/components/ListEditable/styles.module.scss
@@ -1,7 +1,16 @@
+@import "$components/Capabilities/styles/variables.scss";
+
 .listContainer {
   padding-right: 10px;
 }
 
 .listEditableContainer {
   width: 50%;
+}
+
+.input {
+  width: $width;
+  height: $height;
+  font-size: $font-size;
+  margin-bottom: $margin-bottom;
 }
